extract config in a config file
