+++ src/run.js
@@ -6,8 +6,7 @@ const steps = require('./steps');
 const Sentry = require('@sentry/node');
 const initSentry = require('./sentry-init');
 
-const extractConfigurationFromEnvironment = require ('./extract-configuration-from-environment');
-const configuration = extractConfigurationFromEnvironment();
+const { configuration } = require('./config');
 const TIMEOUT = 2000;
 
 async function main() {
