+++ src/schedule-incremental-replication.js
@@ -6,8 +6,7 @@ const logger = require('./logger');
 const Sentry = require('@sentry/node');
 const initSentry = require('./sentry-init');
 
-const extractConfigurationFromEnvironment = require ('./extract-configuration-from-environment');
-const configuration = extractConfigurationFromEnvironment();
+const { configuration } = require('./config');
 
 new CronJob(configuration.SCHEDULE, async function() {
   try {
